chore: make boostFee u8 and bump to 1.8.0-rc.6

+++ packages/shared/src/vault/executeSwap.ts
@@ -84,7 +84,7 @@ const vaultSwapParametersCodec = Struct({
     minPriceX128: u256,
   }),
   dcaParams: Option(Struct({ numberOfChunks: u32, chunkIntervalBlocks: u32 })),
-  boostFee: u16,
+  boostFee: u8,
   brokerFees: Struct({ account: TsBytes(32), commissionBps: u16 }),
   affiliateFees: Vector(Struct({ account: u8, commissionBps: u8 })),
 });
